fix Connected to undefined message (#284)
* fix Connected to undefined message

ENG-5604

* add connection message to all commands

ENG-5630
This will display the endpoint and database being connected to when ever a CLI command is executed.

* fix tests

* allow cmds to opt out of connection info

For eval, we currently output json.  Entering a log line can make the output invalid json and could possibly break tooling people have around it.  This allows commands to opt out of the connection info logging.