don't blow up for invalid config
This will make it such that commands emit warnings when invalid endpoints are detected in the fauna shell config.  Users will still be able to use the shell if this is the case, all invalid endpoints will be ignored.
As it currently stands, any attempt to update the fauna shell config file will fail if there are invalid endpoints as we would remove them in that scenario.