add schema tests to the shell pipeline
I validated that the task works as expected with fly execute.  Once I
verify it is working as expected in the pipeline I'll add it as a passed
constraint to the publish job.  I'll also move the integ tests to the
test job, that way we will always auto run the tests and publish can be
executed with the latest version that passed the tests (or any selected
version that passed the tests)