Remove punycode deprecation warnings
This project's package.json engines field declared support all the way
back to node 10, but we've taken dependencies that are incompatible with
anything older than node 18! After updating the engines field, I did
some digging and realized that native nodeJS fetch landed in v18, so we
have an opportunity here to formalize our support boundary as node 18
and throw out the several fetch implementations we had lying around (one
of which was responsible for the punycode deprecation warnings).

Also includes a .nvmrc to pin to v18 so we don't accidentally include
libraries that require v19+ in the future.

Previous error installing on node v17:
> yarn install
error mute-stream@1.0.0: The engine "node" is incompatible with this module. Expected version "^14.17.0 || ^16.13.0 || >=18.0.0". Got "17.9.1"
error Found incompatible module.

Previous error running on node v17:
> ./bin/run eval "Database.all()"
 ›   Error: fetch is not defined: