lint and prettify on PR push/change
this change updates the linting github action to commit the changes that
prettier causes. this will keep the codebase consistently prettified
without requiring client-side git hooks. it could also commit linting
changes (by running `npm run lint --fix`), but this PR does not enable
that out of an abundance of caution for the changes the linter might
introduce.